Update gemini platform and script
This update comes after the validation of PLIC functionality.
Andes_A45 has been included to the compilatio of cores, so explicitly
including it is no longer needed. Also, the gemini.resc can be executed
directly from the scripts/single-node folder and it is fully recommended
to keen the development from it. Only path to be updated is to properly
point to the .elf file