Update gemini .repl and .resc files (#1)
* Delete previous Infrastructure submodule

* Change renode-infrastructure submodule

In order to integrate our Andes A45 model we need to replace the
renode-infrastructure submodule to point to our fork

* Add working files for gemini

* Change renode script to use Andes_A45.cs

* Update gemini platform and script

This update comes after the validation of PLIC functionality.
Andes_A45 has been included to the compilatio of cores, so explicitly
including it is no longer needed. Also, the gemini.resc can be executed
directly from the scripts/single-node folder and it is fully recommended
to keen the development from it. Only path to be updated is to properly
point to the .elf file

* Update pointer to renode-infraestructure repo

* Fix size of SRAM in platform description